[1] Реализация компонента тегов
