+++ src/app/pages/articles/components/article-view/article-view.component.html
@@ -52,3 +52,10 @@
         readonly="true"
         autosize="true"></in-textarea>
 </div>
+
+<div class="article-row">
+    <app-tag-list
+        [tags]="tags"
+        [tagsOptions]="tagsOptions"
+    ></app-tag-list>
+</div>
