+++ src/app/pages/articles/components/article-view/article-view.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { IArticle } from 'src/app/shared/models/article.interfaces';
+import { IArticle, ITagOptions } from 'src/app/shared/models/article.interfaces';
 
 @Component({
   selector: 'app-article-view',
@@ -9,6 +9,8 @@ import { IArticle } from 'src/app/shared/models/article.interfaces';
 export class ArticleViewComponent {
   @Input() currentArticle: IArticle | null = null;
   @Input() difficultyName = '';
+  @Input() tags: string[] = [];
+  @Input() tagsOptions: ITagOptions[] = [];
 
   @Output() edit = new EventEmitter();
   @Output() delete = new EventEmitter();
